webgpu: export webgpu utilities
