+++ tfjs-backend-webgpu/src/index.ts
@@ -15,45 +15,5 @@
  * =============================================================================
  */
 
-import './flags_webgpu';
+export * from './base';
 import './register_all_kernels';
-
-import {env, registerBackend} from '@tensorflow/tfjs-core';
-
-import {WebGPUBackend} from './backend_webgpu';
-import * as webgpu from './webgpu';
-import {isWebGPUSupported} from './webgpu_util';
-
-if (isWebGPUSupported()) {
-  registerBackend('webgpu', async () => {
-    // Remove it once we figure out how to correctly read the tensor data
-    // before the tensor is disposed in profiling mode.
-    env().set('CHECK_COMPUTATION_FOR_ERRORS', false);
-
-    const gpuDescriptor: GPURequestAdapterOptions = {
-      powerPreference: env().get('WEBGPU_USE_LOW_POWER_GPU') ?
-          'low-power' :
-          'high-performance'
-    };
-
-    const adapter = await navigator.gpu.requestAdapter(gpuDescriptor);
-    const adapterLimits = adapter.limits;
-    const deviceDescriptor: GPUDeviceDescriptor = {};
-    const supportTimeQuery = adapter.features.has('timestamp-query');
-    deviceDescriptor.requiredLimits = {
-      'maxComputeWorkgroupStorageSize':
-          adapterLimits.maxComputeWorkgroupStorageSize,
-      'maxComputeWorkgroupsPerDimension':
-          adapterLimits.maxComputeWorkgroupsPerDimension,
-      'maxStorageBufferBindingSize': adapterLimits.maxStorageBufferBindingSize,
-    };
-
-    if (supportTimeQuery) {
-      deviceDescriptor.requiredFeatures = ['timestamp-query' as const];
-    }
-    const device: GPUDevice = await adapter.requestDevice(deviceDescriptor);
-    return new WebGPUBackend(device);
-  }, 3 /*priority*/);
-}
-
-export {webgpu};
